OSIS-6256 : AssessmentsHome & OutsidePeriod
